requirements.txt
pandas
Pillow
scikit-learn
torch
torchvision
matplotlib